Updated Syntax for Ehancement
As per feedback, revert "smart" components to Class syntax, as future
enahancements may require component methods or lifecycle hooks. May be a
more flexible solution towards smart/dumb architectures.

Created Message component so that it can be "dumber", and allow
MessageList component to worry about iteration/mapping.

Changes to be committed:

-	new file:   app/javascript/bundles/App/components/Message.js
-	deleted:    app/javascript/bundles/App/components/MessageList.js
-	modified:   app/javascript/bundles/App/components/TextDisplay.js
-	modified:   app/javascript/bundles/App/containers/MessageInput.js
-	new file:   app/javascript/bundles/App/containers/MessageList.js
-	deleted:    app/javascript/bundles/App/containers/Messages.js
-	modified:   app/javascript/bundles/App/containers/SomethingSmall.js
